Split tests into logical groups
