+++ src/test/java/io/github/ust/mico/kafkafaasconnector/FaasControllerTests.java
@@ -1,0 +1,69 @@
+package io.github.ust.mico.kafkafaasconnector;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import io.cloudevents.json.Json;
+import io.github.ust.mico.kafkafaasconnector.exception.MicoCloudEventException;
+import io.github.ust.mico.kafkafaasconnector.kafka.MicoCloudEventImpl;
+import io.github.ust.mico.kafkafaasconnector.messageprocessing.FaasController;
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.kafka.test.EmbeddedKafkaBroker;
+import org.springframework.kafka.test.rule.EmbeddedKafkaRule;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.ArrayList;
+
+import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
+@EnableAutoConfiguration
+@AutoConfigureMockMvc
+@ActiveProfiles("testing")
+public class FaasControllerTests {
+
+    /**
+     * We need the embedded Kafka to successfully create the context.
+     */
+    private final EmbeddedKafkaBroker embeddedKafka = broker.getEmbeddedKafka();
+
+    // https://docs.spring.io/spring-kafka/docs/2.2.6.RELEASE/reference/html/#kafka-testing-junit4-class-rule
+    @ClassRule
+    public static EmbeddedKafkaRule broker = new EmbeddedKafkaRule(1, false);
+
+    @Autowired
+    FaasController faasController;
+
+    @Test
+    public void parseEmptyFunctionResult() throws MicoCloudEventException {
+        ArrayList<MicoCloudEventImpl<JsonNode>> result = this.faasController.parseFunctionResult("[]", null);
+        assertNotNull(result);
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void parseFunctionResult() throws MicoCloudEventException {
+        MicoCloudEventImpl<JsonNode> cloudEvent1 = CloudEventTestUtils.basicCloudEvent("CloudEvent1");
+        MicoCloudEventImpl<JsonNode> cloudEvent2 = CloudEventTestUtils.basicCloudEvent("CloudEvent2");
+        ArrayList<MicoCloudEventImpl<JsonNode>> input = new ArrayList<>();
+        input.add(cloudEvent1);
+        input.add(cloudEvent2);
+        String functionInput = Json.encode(input);
+        ArrayList<MicoCloudEventImpl<JsonNode>> result = this.faasController.parseFunctionResult(functionInput, null);
+        assertNotNull(result);
+        assertEquals(2, result.size());
+        assertEquals(result.get(0).getId(), cloudEvent1.getId());
+        assertEquals(result.get(0).getSource(), cloudEvent1.getSource());
+        assertEquals(result.get(0).getType(), cloudEvent1.getType());
+        assertTrue(result.get(0).getTime().get().isEqual(cloudEvent1.getTime().get()));
+        assertEquals(result.get(1).getId(), cloudEvent2.getId());
+        assertEquals(result.get(0).getRoutingSlip(), cloudEvent2.getRoutingSlip());
+    }
+}
