+++ src/test/java/io/github/ust/mico/kafkafaasconnector/KafkaMessageSenderTests.java
@@ -1,0 +1,83 @@
+package io.github.ust.mico.kafkafaasconnector;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import io.github.ust.mico.kafkafaasconnector.configuration.KafkaConfig;
+import io.github.ust.mico.kafkafaasconnector.kafka.MicoCloudEventImpl;
+import io.github.ust.mico.kafkafaasconnector.messageprocessing.KafkaMessageSender;
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.test.EmbeddedKafkaBroker;
+import org.springframework.kafka.test.rule.EmbeddedKafkaRule;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
+@EnableAutoConfiguration
+@AutoConfigureMockMvc
+@ActiveProfiles("testing")
+public class KafkaMessageSenderTests {
+
+    @Autowired
+    KafkaMessageSender kafkaMessageSender;
+
+    /**
+     * We need the embedded Kafka to successfully create the context.
+     */
+    private final EmbeddedKafkaBroker embeddedKafka = broker.getEmbeddedKafka();
+
+    // https://docs.spring.io/spring-kafka/docs/2.2.6.RELEASE/reference/html/#kafka-testing-junit4-class-rule
+    @ClassRule
+    public static EmbeddedKafkaRule broker = new EmbeddedKafkaRule(1, false);
+
+
+    /**
+     * Tests if the need to filter out a message with "isTestMessage = true" and a "FilterOutBeforeTopic = message destination"
+     * is correctly recognized.
+     */
+    @Test
+    public void testFilterOutCheck() {
+        MicoCloudEventImpl<JsonNode> cloudEventSimple = CloudEventTestUtils.basicCloudEventWithRandomId();
+        String testFilterTopic = "TestFilterTopic";
+        cloudEventSimple.setFilterOutBeforeTopic(testFilterTopic);
+        cloudEventSimple.setIsTestMessage(true);
+
+        assertTrue("The message should be filtered out", kafkaMessageSender.isTestMessageCompleted(cloudEventSimple, testFilterTopic));
+    }
+
+    /**
+     * Only "FilterOutBeforeTopic = message destination" is not enough to filter a message out. It needs to be a test message.
+     */
+    @Test
+    public void testNotFilterOutCheck() {
+        MicoCloudEventImpl<JsonNode> cloudEventSimple = CloudEventTestUtils.basicCloudEventWithRandomId();
+        String testFilterTopic = "TestFilterTopic";
+        cloudEventSimple.setFilterOutBeforeTopic(testFilterTopic);
+
+        assertFalse("The message not should be filtered out, because it is not a test message", kafkaMessageSender.isTestMessageCompleted(cloudEventSimple, testFilterTopic));
+    }
+
+    /**
+     * Tests if a test message with "FilterOutBeforeTopic != message destination" will wrongly be filtered out.
+     */
+    @Test
+    public void testNotFilterOutCheckDifferentTopics() {
+        MicoCloudEventImpl<JsonNode> cloudEventSimple = CloudEventTestUtils.basicCloudEventWithRandomId();
+        String testFilterTopic = "TestFilterTopic";
+        cloudEventSimple.setFilterOutBeforeTopic(testFilterTopic);
+        cloudEventSimple.setIsTestMessage(true);
+
+        assertFalse("The message not should be filtered out, because it has not reached the filter out topic", kafkaMessageSender.isTestMessageCompleted(cloudEventSimple, testFilterTopic + "Difference"));
+    }
+}
